Move typedef to top of files
