+++ src/game/backend.js
@@ -2,7 +2,6 @@ import {isDungeonCompleted} from './utils-state.js'
 
 const apiUrl = 'https://api.slaytheweb.cards/api/runs'
 // const apiUrl = 'http://localhost:3000/api/runs'
-//
 
 /**
  * @typedef {object} Run
