+++ src/game/new-game.js
@@ -1,15 +1,17 @@
 import actions from './actions.js'
 import ActionManager from './action-manager.js'
 
+/** @typedef {import('./actions.js').State} State */
+
 /**
  * @typedef {object} Game
- * @prop {import('./actions.js').State} state
+ * @prop {State} state
  * @prop {object} actions
  * @prop {Function} enqueue - stores an action in the "future"
  * @prop {Function} dequeue - runs the oldest "future" action, and stores result in the "past"
  * @prop {Function} undo - undoes the last "past" action
  * @prop {{list: Array<{type: string}>}} future
- * @prop {{list: Array<{type: string, state: import('./actions.js').State}>}} past
+ * @prop {{list: Array<{type: string, state: State}>}} past
  */
 
 /**
