+++ src/game/rooms.js
@@ -1,10 +1,8 @@
-import {shuffle} from '../utils.js'
-import {easyMonsters, monsters, elites, bosses} from '../content/dungeon-encounters.js'
-
+/** @typedef {import('./monster.js').MONSTER} MONSTER */
 /**
  * @typedef {object} Room
  * @prop {RoomTypes} type
- * @prop {Array<import('../game/monster.js').MONSTER>} [monsters] - for monster rooms
+ * @prop {Array<MONSTER>} [monsters] - for monster rooms
  * @prop {object} [reward] - the reward given to the player, if any
  * @prop {string} [choice] - for campfire rooms, the choice made by the player
  */
@@ -49,21 +47,4 @@ export function MonsterRoom(...monsters) {
 		type: 'monster',
 		monsters,
 	}
-}
-
-/**
- * Create a room from the node's type
- * @param {string} type
- * @param {number} floor
- * @returns {Room}
- */
-export function decideRoomType(type, floor) {
-	const pickRandomFromObj = (obj) => obj[shuffle(Object.keys(obj))[0]]
-	if (floor === 0) return StartRoom()
-	if (type === 'C') return CampfireRoom()
-	if (type === 'M' && floor < 2) return pickRandomFromObj(easyMonsters)
-	if (type === 'M') return pickRandomFromObj(monsters)
-	if (type === 'E') return pickRandomFromObj(elites)
-	if (type === 'boss') return pickRandomFromObj(bosses)
-	throw new Error(`Could not match node type "${type}" with a dungeon room`)
 }
